Move IdentityAdapter and Action creation static methods and classes out of MessageCompose
